requirements.txt
# AI Model QA & Hallucination Detection Framework
# Core Dependencies

# OpenAI API integration
openai>=1.12.0

# HuggingFace integration
transformers>=4.38.0
huggingface-hub>=0.20.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0

# NLP utilities
nltk>=3.8.1
rouge-score>=0.1.2
scikit-learn>=1.4.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
